pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mini-datahub"
version = "0.55.0-beta"
description = "A local-first TUI for managing datasets with consistent metadata"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "you@example.com"}
]
keywords = ["data", "catalog", "tui", "metadata", "search"]
classifiers = [
    "Development Status :: 5 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "textual>=0.47.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "jsonschema>=4.20.0",
    "requests>=2.31.0",
    "pyperclip>=1.8.2",
    "keyring>=24.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
    "types-pyyaml",
    "types-requests",
]

[project.scripts]
mini-datahub = "mini_datahub.cli.main:main"
hei-datahub = "mini_datahub.cli.main:main"

[project.urls]
Homepage = "https://github.com/0xpix/Hei-DataHub"
Repository = "https://github.com/0xpix/Hei-DataHub"

[tool.setuptools.packages.find]
where = ["src"]
include = ["mini_datahub*"]

[tool.setuptools.package-data]
mini_datahub = ["infra/sql/*.sql"]

[tool.black]
line-length = 100
target-version = ['py39']

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
